requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
gradio>=4.0.0
twilio>=8.10.0
python-dotenv>=1.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
